requirements.txt
streamlit
unifyai